quick update to the UPDATE
Changed the bat from the old one (i forgot to upload the right one lol) to the new one that has a new feature checking for the python request library. And also added the folder that i forgot to add called images where all of the images will be stored.